pyproject.toml
[tool.poetry]
name = "mybot"
version = "0.1.0"
description = "mybot"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.0.0-beta.2"
nonebot-adapter-onebot = "^2.0.0-beta.1"
nonebot-plugin-apscheduler = "^0.1.2"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = ["nonebot_plugin_apscheduler"]
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
default = true
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"